Moved detouring initializer to header
Switched to singleton on Hook class
