+++ helpers.hpp
@@ -49,6 +49,9 @@
 
 namespace Detouring
 {
+	// Initializes the Detouring library, deinitializes automatically at module unloading
+	void Initialize( );
+
 	namespace MemoryProtection
 	{
 		enum : int32_t
