+++ hook.hpp
@@ -39,13 +39,6 @@
 
 namespace Detouring
 {
-	class Initializer
-	{
-	public:
-		Initializer( );
-		~Initializer( );
-	};
-
 	class Hook
 	{
 	public:
